Give an appropriate error message if ssh sends an argument to askpass.
This usually occurs when it has a "Are you sure you want to continue
connecting (yes/no)?" message and it overloads askpass to try to interact
with the user.

Note that a lazy way to work around this problem is to send ssh "-o
StrictHostKeyChecking=no" or even "-o UserKnownHostsFile=/dev/null", too.